Renamed the classes - prefixed with "Ditto"
